+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffBuildInFnFuncSigCallMixin.kt
@@ -1,0 +1,23 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
+
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffBuildInFnFuncSigCall
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.HuffBuildInFnFuncSigCallElement
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffBuildInFnFuncSigCallReference
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffReference
+import com.intellij.lang.ASTNode
+import com.intellij.psi.PsiElement
+
+abstract class HuffBuildInFnFuncSigCallMixin(node: ASTNode) :
+  HuffNamedElementImpl(node), HuffBuildInFnFuncSigCall, HuffBuildInFnFuncSigCallElement {
+
+  override val referenceNameElement: PsiElement
+    get() = this.identifier ?: this
+
+  override val referenceName: String
+    get() = referenceNameElement.text
+
+  override fun getName(): String? = referenceName
+
+  override fun getReference(): HuffReference = HuffBuildInFnFuncSigCallReference(this as HuffBuildInFnFuncSigCall)
+}